arangoctl: initial checkin with project setup
-- and setup of first apply command
-- setup pkg folder
-- provide implementation for collection and index creation

Based on cluster config and passed config parameter, the tool will
idempotently create collection and it indexes.

If the collection already exists, it will verify that the collection
type matches: collection and edge.

Otherwise, it checks all indexes present in system and compares them
with what is asked in config file and will delete that are not present.
Applying the same is idempotent.